+++ tests/functions.test.js
@@ -1,6 +1,7 @@
 // Require
 const fs = require('fs');
-const units = require('../utils/units');
+const crosstest = require('../utils/crosstest');
+const finder = require('../utils/finder');
 const Bundle = require('../dist/bundle');
 const Lime = require('../src/lime');
 
@@ -12,9 +13,9 @@ let count = 0;
 // Counting test
 describe('Function counting test', () => {
   fs.readdirSync('src/functions').forEach((folder) => {
-    units(`src/functions/${folder}`).forEach((key) => {
+    finder(`src/functions/${folder}`).forEach((key) => {
       test(`${folder}/${key}`, () => {
-        expect(units(`tests/functions/${folder}`, '.json').includes(key)).toBeTruthy();
+        expect(finder(`tests/functions/${folder}`, '.json').includes(key)).toBeTruthy();
       });
     });
   });
@@ -22,23 +23,11 @@ describe('Function counting test', () => {
 
 // Unit tests
 fs.readdirSync('tests/functions').forEach((folder) => {
-  units(`tests/functions/${folder}`, '.json').forEach((key) => {
+  finder(`tests/functions/${folder}`, '.json').forEach((key) => {
     const unit = require(`./functions/${folder}/${key}`);
 
     describe(unit.name, () => {
-      Object.values(unit.process).forEach((obj) => {
-        const len = obj.tests.length;
-        test(`${obj.desc} (${len} test${len === 1 ? '' : 's'})`, () => {
-          [Lime, Bundle].forEach((p) => {
-            for (let item in obj.tests) {
-              const lime = (p)({ testMode: true, ...unit.config }), l = obj.tests[item].length;
-              obj.tests[item].slice(0, -2).forEach((j) => lime.prompt(j));
-              expect(lime.prompt(obj.tests[item][l - 2])).toStrictEqual(obj.tests[item][l - 1]);
-            }
-          });
-        });
-        count += len;
-      });
+      count += crosstest(test, expect, Lime, Bundle, unit);
     });
   });
 });