+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### v0.1.2
+- Set up prompt and evaluate modes for input
+- Set up crosstesting system and standardized process-based tests
 ### v0.1.1
 - Set up command system in prompt
 - Added commands: `About`, `Help`